restructured project folders & namespaces
