+++ app/Presenter/Error4xxPresenter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace App\Presenters;
+namespace App\Presenter;
 
 use Nette\Application\Request;
 use Nette\Application\UI\Presenter;
