+++ app/Presenter/HomepagePresenter.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace App\Presenters;
+namespace App\Presenter;
 
 use Nette\Application\UI\Presenter;
