+++ app/Presenter/templates/Error/500.phtml
