+++ app/config/common.neon
@@ -3,18 +3,19 @@ search:
 	exclude:
 		classes:
 			- App\Bootstrap
-			- App\RouterFactory
+			- App\Router\RouterFactory
 
 
 application:
 	errorPresenter: Error
 	mapping:
-		*: App\*Module\Presenters\*Presenter
+		*: App\*Module\Presenter\*Presenter
 
 
 session:
 	expiration: 14 days
+	save_path: %tempDir%/../sessions
 
 
 services:
-	router: App\RouterFactory::createRouter
+	router: App\Router\RouterFactory::createRouter
